src/test/java/marc/springframework/spring6restmvc/BeerCsvServiceImplTest.java
package marc.springframework.spring6restmvc;

import marc.springframework.spring6restmvc.model.BeerCSVRecord;
import marc.springframework.spring6restmvc.services.BeerCsvService;
import marc.springframework.spring6restmvc.services.BeerCsvServiceImpl;
import org.junit.jupiter.api.Test;
import org.springframework.util.ResourceUtils;

import java.io.File;
import java.io.FileNotFoundException;
import java.util.List;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

public class BeerCsvServiceImplTest
{
    BeerCsvService beerCsvService = new BeerCsvServiceImpl();

    @Test
    void convertCSV() throws FileNotFoundException
    {
        File file = ResourceUtils.getFile("classpath:csvdata/beers.csv");

        List<BeerCSVRecord> recs = beerCsvService.convertCSV(file);

        System.out.println(recs.size());

        assertThat(recs.size()).isGreaterThan(0);
    }
}
